26305 - Correção Querys planejamento anual
